Abstract out the editbar
